Update OSIsoft Adapter for Modbus TCP data source configuration.md
Updated the data source configuration to reflect changes made in the Atlantis program. These include allowing for multiple data sources, as shown by the new "Sources" array parameter, the addition of the SimultaneousRequests parameter, updates to stream id prefixing with the new StreamIdPrefix and DefaultStreamIdPattern parameters, and the update of "time" parameters to use a TimeSpan string value.